Copy archive first to tempfile in same directory
Leaving a partially-copied archive with the relevant name in the storage dir creates a problem since there is an archive present, so no new virtualenv is built, but it can't be used. We should first write to a temporary file in the same directory, then do `os.rename()` which is guaranteed to be atomic.